services/inference_service/requirements.txt
mlflow==2.9.2
pandas
scikit-learn
numpy
confluent-kafka
kafka-python